Here is a possible commit message:
"Updated language files for Italian translation (#1234)

This commit updates the language files for the Italian translation, including changes to navigation menus and page headings. The updated files include improvements to the session.php and session.php files.

Changes include:
- Updated navigation menu items
- Improved page headings
- Added new translations for session-related terms

This update is part of our ongoing effort to improve the Italian translation of our application. We hope this will help users in Italy better navigate and understand our application."